refactor(bench): 対局ループを runMatch に抽出(weight-bench と共有)
commit-bench の main() に埋もれていた珠型タスク生成・bridge worker 生成・
ワークスティール並列ループ・WDL/Elo/SPRT/状態表示を scripts/lib/match.ts へ抽出。
後続の weight-bench(ローカルwasm+重み注入)が同じ対局ループを再利用できるようにする。

- commit-bench は recreatePair を渡さない＝エラー伝播の旧挙動を維持(出力同一)
- weight-bench 用に1局隔離(try/catch+worker再生成)を recreatePair 指定時のみ有効化
- createBridgeWorker は evalWeights/loaderPath を引数化(後続段階で使用)
